oboeru: only timestamp review on first review of the session
this mostly reverts the last commit which was based on a misunderstanding
of the problem. the reason we were hitting that branch twice was because
the review timestamp was being updated more than once.

we still add an extra 1s to avoid any rounding error the next time the
card is reviewed